feat: reth-config crate
