+++ bin/reth/src/args/network_args.rs
@@ -2,10 +2,10 @@
 
 use crate::version::P2P_CLIENT_VERSION;
 use clap::Args;
+use reth_config::Config;
 use reth_net_nat::NatResolver;
 use reth_network::{HelloMessage, NetworkConfigBuilder};
 use reth_primitives::{mainnet_nodes, ChainSpec, NodeRecord};
-use reth_staged_sync::Config;
 use secp256k1::SecretKey;
 use std::{path::PathBuf, sync::Arc};
