+++ bin/reth/src/config.rs
@@ -3,7 +3,7 @@ use std::path::PathBuf;
 
 use clap::Parser;
 use eyre::{bail, WrapErr};
-use reth_staged_sync::Config;
+use reth_config::Config;
 
 /// `reth config` command
 #[derive(Debug, Parser)]
