+++ bin/reth/src/stage/run.rs
@@ -9,13 +9,11 @@ use crate::{
 };
 use clap::Parser;
 use reth_beacon_consensus::BeaconConsensus;
+use reth_config::Config;
 use reth_downloaders::bodies::bodies::BodiesDownloaderBuilder;
 use reth_primitives::{ChainSpec, StageCheckpoint};
 use reth_provider::{ShareableDatabase, Transaction};
-use reth_staged_sync::{
-    utils::{chainspec::chain_spec_value_parser, init::init_db},
-    Config,
-};
+use reth_staged_sync::utils::{chainspec::chain_spec_value_parser, init::init_db};
 use reth_stages::{
     stages::{
         BodyStage, ExecutionStage, ExecutionStageThresholds, MerkleStage, SenderRecoveryStage,
