+++ crates/config/src/lib.rs
@@ -1,0 +1,11 @@
+#![warn(missing_docs, unreachable_pub)]
+#![deny(unused_must_use, rust_2018_idioms)]
+#![doc(test(
+    no_crate_inject,
+    attr(deny(warnings, rust_2018_idioms), allow(dead_code, unused_variables))
+))]
+
+//! Standalone crate for Reth config types
+
+pub mod config;
+pub use config::Config;
